+++ src/fugit/pwm.rs
@@ -1,6 +1,7 @@
-use super::{General, Timer};
+use super::{Channel, Instance, Ocm, Timer, WithPwm};
 use crate::rcc::Clocks;
-use core::mem::MaybeUninit;
+use core::marker::PhantomData;
+use core::ops::{Deref, DerefMut};
 use fugit::TimerDurationU32;
 
 pub use crate::pwm::{Pins, PwmChannel};
@@ -8,7 +9,7 @@ pub use crate::timer::{CPin, C1, C2, C3, C4};
 
 pub trait PwmExt<P, PINS>
 where
-    Self: Sized,
+    Self: Sized + Instance + WithPwm,
     PINS: Pins<Self, P>,
 {
     fn pwm<const FREQ: u32>(
@@ -16,283 +17,165 @@ where
         clocks: &Clocks,
         pins: PINS,
         time: TimerDurationU32<FREQ>,
-    ) -> PINS::Channels;
+    ) -> Pwm<Self, P, PINS, FREQ>;
 
     fn pwm_us(
         self,
         clocks: &Clocks,
         pins: PINS,
         time: TimerDurationU32<1_000_000>,
-    ) -> PINS::Channels {
+    ) -> Pwm<Self, P, PINS, 1_000_000> {
         self.pwm::<1_000_000>(clocks, pins, time)
     }
 }
 
-macro_rules! brk {
-    (TIM1, $tim:ident) => {
-        $tim.bdtr.modify(|_, w| w.aoe().set_bit());
-    };
-    (TIM8, $tim:ident) => {
-        $tim.bdtr.modify(|_, w| w.aoe().set_bit());
-    };
-    ($_other:ident, $_tim:ident) => {};
+impl<TIM, P, PINS> PwmExt<P, PINS> for TIM
+where
+    Self: Sized + Instance + WithPwm,
+    PINS: Pins<Self, P>,
+{
+    fn pwm<const FREQ: u32>(
+        self,
+        clocks: &Clocks,
+        pins: PINS,
+        time: TimerDurationU32<FREQ>,
+    ) -> Pwm<TIM, P, PINS, FREQ> {
+        Timer::<Self, FREQ>::new(self, clocks).pwm(pins, time)
+    }
 }
 
-macro_rules! pwm_all_channels {
-    ($($TIMX:ident,)+) => {
-        $(
-            impl<P, PINS> PwmExt<P, PINS> for crate::pac::$TIMX where
-                Self: Sized,
-                PINS: Pins<Self, P>
-            {
-                fn pwm<const FREQ: u32>(self, clocks: &Clocks, pins: PINS, time: TimerDurationU32<FREQ>) -> PINS::Channels {
-                    Timer::<Self, FREQ>::new(self, clocks).pwm(pins, time)
-                }
-            }
-
-            impl<const FREQ: u32> Timer<crate::pac::$TIMX, FREQ> {
-                pub fn pwm<P, PINS>(mut self, _pins: PINS, time: TimerDurationU32<FREQ>) -> PINS::Channels
-                where
-                    PINS: Pins<crate::pac::$TIMX, P>,
-                {
-                    if PINS::C1 {
-                        self.tim.ccmr1_output()
-                            .modify(|_, w| w.oc1pe().set_bit().oc1m().pwm_mode1() );
-                    }
-                    if PINS::C2 {
-                        self.tim.ccmr1_output()
-                            .modify(|_, w| w.oc2pe().set_bit().oc2m().pwm_mode1() );
-                    }
-                    if PINS::C3 {
-                        self.tim.ccmr2_output()
-                            .modify(|_, w| w.oc3pe().set_bit().oc3m().pwm_mode1() );
-                    }
-                    if PINS::C4 {
-                        self.tim.ccmr2_output()
-                            .modify(|_, w| w.oc4pe().set_bit().oc4m().pwm_mode1() );
-                    }
-
-                    // The reference manual is a bit ambiguous about when enabling this bit is really
-                    // necessary, but since we MUST enable the preload for the output channels then we
-                    // might as well enable for the auto-reload too
-                    self.tim.cr1.modify(|_, w| w.arpe().set_bit());
-
-                    self.tim.set_auto_reload(time.ticks() - 1).unwrap();
-
-                    // Trigger update event to load the registers
-                    self.tim.trigger_update();
-
-                    let _tim = &self.tim;
-                    brk!($TIMX, _tim);
-                    self.tim.cr1.write(|w|
-                        w.cms()
-                            .bits(0b00)
-                            .dir()
-                            .clear_bit()
-                            .opm()
-                            .clear_bit()
-                            .cen()
-                            .set_bit()
-                    );
-                    //NOTE(unsafe) `PINS::Channels` is a ZST
-                    unsafe { MaybeUninit::uninit().assume_init() }
-                }
-            }
-        )+
-    };
+pub struct Pwm<TIM, P, PINS, const FREQ: u32>
+where
+    TIM: Instance + WithPwm,
+    PINS: Pins<TIM, P>,
+{
+    timer: Timer<TIM, FREQ>,
+    _pins: PhantomData<(P, PINS)>,
 }
 
-macro_rules! pwm_2_channels {
-    ($($TIMX:ident,)+) => {
-        $(
-            impl<P, PINS> PwmExt<P, PINS> for crate::pac::$TIMX where
-                Self: Sized,
-                PINS: Pins<Self, P>
-            {
-                fn pwm<const FREQ: u32>(self, clocks: &Clocks, pins: PINS, time: TimerDurationU32<FREQ>) -> PINS::Channels {
-                    Timer::<Self, FREQ>::new(self, clocks).pwm(pins, time)
-                }
-            }
+impl<TIM, P, PINS, const FREQ: u32> Pwm<TIM, P, PINS, FREQ>
+where
+    TIM: Instance + WithPwm,
+    PINS: Pins<TIM, P>,
+{
+    pub fn split(self) -> PINS::Channels {
+        PINS::split()
+    }
 
-            impl<const FREQ: u32> Timer<crate::pac::$TIMX, FREQ> {
-                pub fn pwm<P, PINS>(mut self, _pins: PINS, time: TimerDurationU32<FREQ>) -> PINS::Channels
-                where
-                    PINS: Pins<crate::pac::$TIMX, P>,
-                {
-                    if PINS::C1 {
-                        self.tim.ccmr1_output().modify(|_, w| w.oc1pe().set_bit().oc1m().pwm_mode1());
-
-                    }
-                    if PINS::C2 {
-                        self.tim.ccmr1_output().modify(|_, w| w.oc2pe().set_bit().oc2m().pwm_mode1());
-
-                    }
-
-                    // The reference manual is a bit ambiguous about when enabling this bit is really
-                    // necessary, but since we MUST enable the preload for the output channels then we
-                    // might as well enable for the auto-reload too
-                    self.tim.cr1.modify(|_, w| w.arpe().set_bit());
-
-                    self.tim.set_auto_reload(time.ticks() - 1).unwrap();
-
-                    // Trigger update event to load the registers
-                    self.tim.trigger_update();
-
-                    self.tim.cr1.write(|w|
-                        w.opm()
-                            .clear_bit()
-                            .cen()
-                            .set_bit()
-                    );
-                    //NOTE(unsafe) `PINS::Channels` is a ZST
-                    unsafe { MaybeUninit::uninit().assume_init() }
-                }
-            }
-        )+
-    };
+    pub fn release(mut self) -> Timer<TIM, FREQ> {
+        // stop counter
+        self.tim.cr1_reset();
+        self.timer
+    }
 }
 
-macro_rules! pwm_1_channel {
-    ($($TIMX:ident,)+) => {
-        $(
-            impl<P, PINS> PwmExt<P, PINS> for crate::pac::$TIMX where
-                Self: Sized,
-                PINS: Pins<Self, P>
-            {
-                fn pwm<const FREQ: u32>(self, clocks: &Clocks, pins: PINS, time: TimerDurationU32<FREQ>) -> PINS::Channels {
-                    Timer::<Self, FREQ>::new(self, clocks).pwm(pins, time)
-                }
-            }
-
-            impl<const FREQ: u32> Timer<crate::pac::$TIMX,FREQ> {
-                pub fn pwm<P, PINS>(mut self, _pins: PINS, time: TimerDurationU32<FREQ>) -> PINS::Channels
-                where
-                    PINS: Pins<crate::pac::$TIMX, P>,
-                {
-                    if PINS::C1 {
-                        self.tim.ccmr1_output()
-                            .modify(|_, w| w.oc1pe().set_bit().oc1m().pwm_mode1());
-
-                    }
-
-                    // The reference manual is a bit ambiguous about when enabling this bit is really
-                    // necessary, but since we MUST enable the preload for the output channels then we
-                    // might as well enable for the auto-reload too
-                    self.tim.cr1.modify(|_, w| w.arpe().set_bit());
-
-                    self.tim.set_auto_reload(time.ticks() - 1).unwrap();
-
-                    // Trigger update event to load the registers
-                    self.tim.trigger_update();
-
-                    self.tim.cr1.write(|w|
-                        w.cen()
-                            .set_bit()
-                    );
-                    //NOTE(unsafe) `PINS::Channels` is a ZST
-                    unsafe { MaybeUninit::uninit().assume_init() }
-                }
-            }
-        )+
-    };
+impl<TIM, P, PINS, const FREQ: u32> Deref for Pwm<TIM, P, PINS, FREQ>
+where
+    TIM: Instance + WithPwm,
+    PINS: Pins<TIM, P>,
+{
+    type Target = Timer<TIM, FREQ>;
+    fn deref(&self) -> &Self::Target {
+        &self.timer
+    }
 }
 
-pwm_all_channels!(TIM1, TIM5,);
+impl<TIM, P, PINS, const FREQ: u32> DerefMut for Pwm<TIM, P, PINS, FREQ>
+where
+    TIM: Instance + WithPwm,
+    PINS: Pins<TIM, P>,
+{
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.timer
+    }
+}
 
-pwm_2_channels!(TIM9,);
+impl<TIM: Instance + WithPwm, const FREQ: u32> Timer<TIM, FREQ> {
+    pub fn pwm<P, PINS>(
+        mut self,
+        _pins: PINS,
+        time: TimerDurationU32<FREQ>,
+    ) -> Pwm<TIM, P, PINS, FREQ>
+    where
+        PINS: Pins<TIM, P>,
+    {
+        if PINS::C1 {
+            self.tim
+                .preload_output_channel_in_mode(Channel::C1, Ocm::PwmMode1);
+        }
+        if PINS::C2 && TIM::CH_NUMBER > 1 {
+            self.tim
+                .preload_output_channel_in_mode(Channel::C2, Ocm::PwmMode1);
+        }
+        if PINS::C3 && TIM::CH_NUMBER > 2 {
+            self.tim
+                .preload_output_channel_in_mode(Channel::C3, Ocm::PwmMode1);
+        }
+        if PINS::C4 && TIM::CH_NUMBER > 3 {
+            self.tim
+                .preload_output_channel_in_mode(Channel::C4, Ocm::PwmMode1);
+        }
 
-pwm_1_channel!(TIM11,);
+        // The reference manual is a bit ambiguous about when enabling this bit is really
+        // necessary, but since we MUST enable the preload for the output channels then we
+        // might as well enable for the auto-reload too
+        self.tim.enable_preload(true);
 
-#[cfg(any(
-    feature = "stm32f401",
-    feature = "stm32f405",
-    feature = "stm32f407",
-    feature = "stm32f411",
-    feature = "stm32f412",
-    feature = "stm32f413",
-    feature = "stm32f415",
-    feature = "stm32f417",
-    feature = "stm32f423",
-    feature = "stm32f427",
-    feature = "stm32f429",
-    feature = "stm32f437",
-    feature = "stm32f439",
-    feature = "stm32f446",
-    feature = "stm32f469",
-    feature = "stm32f479"
-))]
-pwm_all_channels!(TIM2, TIM3, TIM4,);
+        self.tim.set_auto_reload(time.ticks() - 1).unwrap();
 
-#[cfg(any(
-    feature = "stm32f401",
-    feature = "stm32f405",
-    feature = "stm32f407",
-    feature = "stm32f411",
-    feature = "stm32f412",
-    feature = "stm32f413",
-    feature = "stm32f415",
-    feature = "stm32f417",
-    feature = "stm32f423",
-    feature = "stm32f427",
-    feature = "stm32f429",
-    feature = "stm32f437",
-    feature = "stm32f439",
-    feature = "stm32f446",
-    feature = "stm32f469",
-    feature = "stm32f479"
-))]
-pwm_1_channel!(TIM10,);
+        // Trigger update event to load the registers
+        self.tim.trigger_update();
 
-#[cfg(any(
-    feature = "stm32f405",
-    feature = "stm32f407",
-    feature = "stm32f412",
-    feature = "stm32f413",
-    feature = "stm32f415",
-    feature = "stm32f417",
-    feature = "stm32f423",
-    feature = "stm32f427",
-    feature = "stm32f429",
-    feature = "stm32f437",
-    feature = "stm32f439",
-    feature = "stm32f446",
-    feature = "stm32f469",
-    feature = "stm32f479"
-))]
-pwm_all_channels!(TIM8,);
+        self.tim.start_pwm();
 
-#[cfg(any(
-    feature = "stm32f405",
-    feature = "stm32f407",
-    feature = "stm32f412",
-    feature = "stm32f413",
-    feature = "stm32f415",
-    feature = "stm32f417",
-    feature = "stm32f423",
-    feature = "stm32f427",
-    feature = "stm32f429",
-    feature = "stm32f437",
-    feature = "stm32f439",
-    feature = "stm32f446",
-    feature = "stm32f469",
-    feature = "stm32f479"
-))]
-pwm_2_channels!(TIM12,);
+        Pwm {
+            timer: self,
+            _pins: PhantomData,
+        }
+    }
+}
 
-#[cfg(any(
-    feature = "stm32f405",
-    feature = "stm32f407",
-    feature = "stm32f412",
-    feature = "stm32f413",
-    feature = "stm32f415",
-    feature = "stm32f417",
-    feature = "stm32f423",
-    feature = "stm32f427",
-    feature = "stm32f429",
-    feature = "stm32f437",
-    feature = "stm32f439",
-    feature = "stm32f446",
-    feature = "stm32f469",
-    feature = "stm32f479"
-))]
-pwm_1_channel!(TIM13, TIM14,);
+impl<TIM, P, PINS, const FREQ: u32> embedded_hal::Pwm for Pwm<TIM, P, PINS, FREQ>
+where
+    TIM: Instance + WithPwm,
+    PINS: Pins<TIM, P>,
+{
+    type Channel = Channel;
+    type Duty = u16;
+    type Time = TimerDurationU32<FREQ>;
+
+    fn enable(&mut self, channel: Self::Channel) {
+        self.tim.enable_channel(PINS::check_used(channel), true)
+    }
+
+    fn disable(&mut self, channel: Self::Channel) {
+        self.tim.enable_channel(PINS::check_used(channel), false)
+    }
+
+    fn get_duty(&self, channel: Self::Channel) -> Self::Duty {
+        let duty: u32 = self.tim.read_cc_value(PINS::check_used(channel)).into();
+        duty as u16
+    }
+
+    fn set_duty(&mut self, channel: Self::Channel, duty: Self::Duty) {
+        self.tim
+            .set_cc_value(PINS::check_used(channel), duty.into())
+    }
+
+    /// If `0` returned means max_duty is 2^16
+    fn get_max_duty(&self) -> Self::Duty {
+        let arr: u32 = self.tim.read_auto_reload().into();
+        (arr as u16).wrapping_add(1)
+    }
+
+    fn get_period(&self) -> Self::Time {
+        Self::Time::from_ticks(self.tim.read_auto_reload().into() + 1)
+    }
+
+    fn set_period<T>(&mut self, period: T)
+    where
+        T: Into<Self::Time>,
+    {
+        self.tim.set_auto_reload(period.into().ticks() - 1).unwrap();
+    }
+}